+++ src/main/java/com/revature/get_user/User.java
@@ -1,4 +1,4 @@
-package com.revature.models;
+package com.revature.get_user;
 
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBAttribute;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
